can get and parse a *part* of JSON message
copying the other parser, I can get one field.

Can't yet get all fields.

This commit also had an important element in that I figured out that
Model is a function, accepting arguments.  So if I want to modify the
model, I can call

```
Model a [b c]
```

and it will happen, because the function signature is looking for a
Model as the first return value, and I can make that by using Model as a
function as above.